Move mkchromecast.py to bin/mkchromecast, shebang fixes.
Distutils/setup.py runs into potential problems when there is both a
module mkchromecast and a script mkchromecast. Resolve issue by moving
script to new location bin/.

Drop the .py suffix since it is a script and make it 755.

Add sys.path code to make it run both when installed and git-cloned.

The shebang is problematic. /usr/bin/python is wrong on most platforms,
since this is a python3 script. Updating to /usr/bin/python3 which
should be fine in many but not all cases -  the upstream python packages
just installs /usr/bin/python3.6 etc., not python3.

When the shebang fails, invocations like "python3.6 scripts/mkchromecast"
will always work (and should be documented somewhere).

Packaging toolchains typically patches the shebang as required, so this
is less of a problem in downstream packaging.