debian: Update to dh compat v13
It was set to 9 via debian/compat which is marked as a deprecated level.
The level is now being set directly in debian/control via

debhelper-compat (= 13)

Therefore debian/compat is no longer needed and has been updated.

Starting with level 13 dh_missing --fail-missing is the default
thus the override in the debian/rules file has been removed.

See as well: man 7 debhelper