Restores a legacy behavior that is likely important in non-tray mode
In particular, because we don't fully-resolve the backend to an absolute
pathname except in tray mode, with a recent change, we would leave
backend.path as None instead of the unqualified backend name.

The better option is to drop manual backend path handling altogether, but until
we make that change, it makes sense to add this fallback back in.